requirements-dev.txt
pytest>=3.4.2
setuptools>=38.5.2
sphinxcontrib-fulltoc==1.2.0
hypothesis>=4.4.0
nbsphinx>=0.4.2
isort>=4.3.18
black>=19.3b0
jupyter_client
natsort
ipykernel
seaborn
pytest-azurepipelines
pytest-cov
lxml
unyt
pyspark
pre-commit
xarray
pandas-vet
